refactor: put code in a single file again
